+++ mood_music.client/src/api/weatherApi.ts
@@ -8,3 +8,12 @@ export const getWeather = async (city: string) => {
         throw error;
     }
 }
+
+export const getWeatherByLocation = async (lat: number, lon: number) => {
+    try {
+        const response = await axios.get(`${import.meta.env.VITE_BACKEND_URL}/api/weather/location?lat=${lat}&lon=${lon}`);
+        return response.data;
+    } catch (error) {
+        throw error;
+    }
+}